+++ reoptjl/urls.py
@@ -23,6 +23,7 @@ urlpatterns = [
     re_path(r'^invalid_urdb/?$', reoviews.invalid_urdb),
     re_path(r'^schedule_stats/?$', reoviews.schedule_stats),
     re_path(r'^get_existing_chiller_default_cop/?$', views.get_existing_chiller_default_cop),
+    re_path(r'^get_ashp_defaults/?$', views.get_ashp_defaults),
     re_path(r'^summary_by_runuuids/?$', views.summary_by_runuuids),
     re_path(r'^link_run_to_portfolios/?$', views.link_run_uuids_to_portfolio_uuid)
 ]