+++ reoptjl/views.py
@@ -15,7 +15,9 @@ from reoptjl.models import Settings, PVInputs, ElectricStorageInputs, WindInputs
     CoolingLoadOutputs, HeatingLoadOutputs, REoptjlMessageOutputs, HotThermalStorageInputs, HotThermalStorageOutputs,\
     ColdThermalStorageInputs, ColdThermalStorageOutputs, AbsorptionChillerInputs, AbsorptionChillerOutputs,\
     FinancialInputs, FinancialOutputs, UserUnlinkedRuns, BoilerInputs, BoilerOutputs, SteamTurbineInputs, \
-    SteamTurbineOutputs, GHPInputs, GHPOutputs, PortfolioUnlinkedRuns, ProcessHeatLoadInputs
+    SteamTurbineOutputs, GHPInputs, GHPOutputs, ProcessHeatLoadInputs, ElectricHeaterInputs, ElectricHeaterOutputs, \
+    ASHPSpaceHeaterInputs, ASHPSpaceHeaterOutputs, ASHPWaterHeaterInputs, ASHPWaterHeaterOutputs, PortfolioUnlinkedRuns
+
 import os
 import requests
 import numpy as np
@@ -60,6 +62,10 @@ def help(request):
         d["AbsorptionChiller"] = AbsorptionChillerInputs.info_dict(AbsorptionChillerInputs)
         d["SteamTurbine"] = SteamTurbineInputs.info_dict(SteamTurbineInputs)
         d["GHP"] = GHPInputs.info_dict(GHPInputs)
+        d["ElectricHeater"] = ElectricHeaterInputs.info_dict(ElectricHeaterInputs)
+        d["ASHPSpaceHeater"] = ASHPSpaceHeaterInputs.info_dict(ASHPSpaceHeaterInputs)
+        d["ASHPWaterHeater"] = ASHPWaterHeaterInputs.info_dict(ASHPWaterHeaterInputs)
+
         return JsonResponse(d)
 
     except Exception as e:
@@ -105,6 +111,9 @@ def outputs(request):
         d["CHP"] = CHPOutputs.info_dict(CHPOutputs)
         d["AbsorptionChiller"] = AbsorptionChillerOutputs.info_dict(AbsorptionChillerOutputs)
         d["GHP"] = GHPOutputs.info_dict(GHPOutputs)
+        d["ElectricHeater"] = ElectricHeaterOutputs.info_dict(ElectricHeaterOutputs)
+        d["ASHPSpaceHeater"] = ASHPSpaceHeaterOutputs.info_dict(ASHPSpaceHeaterOutputs)
+        d["ASHPWaterHeater"] = ASHPWaterHeaterOutputs.info_dict(ASHPWaterHeaterOutputs)
         d["Messages"] = REoptjlMessageOutputs.info_dict(REoptjlMessageOutputs)
         d["SteamTurbine"] = SteamTurbineOutputs.info_dict(SteamTurbineOutputs)
         return JsonResponse(d)
@@ -227,8 +236,18 @@ def results(request, run_uuid):
 
     try: r["inputs"]["SteamTurbine"] = meta.SteamTurbineInputs.dict
     except: pass
+
     try: r["inputs"]["GHP"] = meta.GHPInputs.dict
     except: pass    
+
+    try: r["inputs"]["ElectricHeater"] = meta.ElectricHeaterInputs.dict
+    except: pass    
+
+    try: r["inputs"]["ASHPSpaceHeater"] = meta.ASHPSpaceHeaterInputs.dict
+    except: pass    
+
+    try: r["inputs"]["ASHPWaterHeater"] = meta.ASHPWaterHeaterInputs.dict
+    except: pass  
 
     try:
         r["outputs"] = dict()
@@ -301,6 +320,12 @@ def results(request, run_uuid):
         try: r["outputs"]["SteamTurbine"] = meta.SteamTurbineOutputs.dict
         except: pass
         try: r["outputs"]["GHP"] = meta.GHPOutputs.dict
+        except: pass
+        try: r["outputs"]["ElectricHeater"] = meta.ElectricHeaterOutputs.dict
+        except: pass    
+        try: r["outputs"]["ASHPSpaceHeater"] = meta.ASHPSpaceHeaterOutputs.dict
+        except: pass  
+        try: r["outputs"]["ASHPWaterHeater"] = meta.ASHPWaterHeaterOutputs.dict
         except: pass
 
         for d in r["outputs"].values():
@@ -440,6 +465,38 @@ def absorption_chiller_defaults(request):
                                                                             tb.format_tb(exc_traceback))
         log.debug(debug_msg)
         return JsonResponse({"Error": "Unexpected error in absorption_chiller_defaults endpoint. Check log for more."}, status=500)
+
+def get_ashp_defaults(request):
+    inputs = {}
+    if request.GET.get("load_served") not in [None, "", []]:
+        inputs["load_served"] = request.GET.get("load_served")
+    else: 
+        return JsonResponse({"Error: Missing input load_served in get_ashp_defualts endpoint."}, status=400)
+    if request.GET.get("force_into_system") not in [None, "", []]:
+        inputs["force_into_system"] = request.GET.get("force_into_system")
+    else: 
+        return JsonResponse({"Error: Missing input force_into_system in get_ashp_defualts endpoint."}, status=400)
+    try:
+        julia_host = os.environ.get('JULIA_HOST', "julia")
+        http_jl_response = requests.get("http://" + julia_host + ":8081/get_ashp_defaults/", json=inputs)
+        response = JsonResponse(
+            http_jl_response.json(),
+            status=http_jl_response.status_code
+        )
+        return response
+
+    except ValueError as e:
+        return JsonResponse({"Error": str(e.args[0])}, status=400)
+
+    except KeyError as e:
+        return JsonResponse({"Error. Missing": str(e.args[0])}, status=400)
+
+    except Exception:
+        exc_type, exc_value, exc_traceback = sys.exc_info()
+        debug_msg = "exc_type: {}; exc_value: {}; exc_traceback: {}".format(exc_type, exc_value.args[0],
+                                                                            tb.format_tb(exc_traceback))
+        log.debug(debug_msg)
+        return JsonResponse({"Error": "Unexpected error in get_ashp_defaults endpoint. Check log for more."}, status=500)
 
 
 def simulated_load(request):
@@ -1189,6 +1246,29 @@ def queryset_for_summary(api_metas,summary_dict:dict):
                     summary_dict[str(m.meta.run_uuid)]['ghp_heating_ton'] = m.size_wwhp_heating_pump_ton
                 summary_dict[str(m.meta.run_uuid)]['ghp_n_bores'] = m.ghpghx_chosen_outputs['number_of_boreholes']
     
+    elecHeater = ElectricHeaterOutputs.objects.filter(meta__run_uuid__in=run_uuids).only(
+            'meta__run_uuid',
+            'size_mmbtu_per_hour'
+    )
+    if len(elecHeater) > 0:
+        for m in elecHeater:
+            summary_dict[str(m.meta.run_uuid)]['electric_heater_mmbtu_per_hour'] = m.size_mmbtu_per_hour
+
+    ashpSpaceHeater = ASHPSpaceHeaterOutputs.objects.filter(meta__run_uuid__in=run_uuids).only(
+            'meta__run_uuid',
+            'size_ton'
+    )
+    if len(ashpSpaceHeater) > 0:
+        for m in ashpSpaceHeater:
+            summary_dict[str(m.meta.run_uuid)]['ASHPSpace_heater_ton'] = m.size_ton
+
+    ashpWaterHeater = ASHPWaterHeaterOutputs.objects.filter(meta__run_uuid__in=run_uuids).only(
+            'meta__run_uuid',
+            'size_ton'
+    )
+    if len(ashpWaterHeater) > 0:
+        for m in ashpWaterHeater:
+            summary_dict[str(m.meta.run_uuid)]['ASHPWater_heater_ton'] = m.size_ton
     hottes = HotThermalStorageOutputs.objects.filter(meta__run_uuid__in=run_uuids).only(
         'meta__run_uuid',
         'size_gal'